requirements.txt
anyio==3.6.2
APScheduler==3.6.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.39.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.11.0
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
schedule==1.2.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tornado==6.3.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wsproto==1.2.0
